feat: add CORS conf, replace hardcoded UUID for real one
